requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
feedparser>=6.0.0
trafilatura>=2.0.0
playwright>=1.49.0
httpx>=0.28.0
apscheduler>=3.11.0,<4.0.0
pydantic-settings>=2.0.0
sentence-transformers>=3.3.0
pgvector>=0.3.0
structlog>=24.4.0
ollama>=0.6.0
sparqlwrapper>=2.0.0
jinja2>=3.1.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.18
listparser>=0.20
beautifulsoup4>=4.12.0
pillow>=11.0.0
redis>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-timeout>=2.3.0
